Add styling to shopping Items
Shopping list displays a grid that is responsive to screen size.  Each shopping item is in a material ui paper element because I liked the box shadow and border radius it comes with.  I used display absolute to place the icons where I wanted them and also placed the text more in the middle.  I added colors from our theme.